Introduce a non-blocking "event loop" to the command processing
These changes change the FakeBus to process successfully submitted
commands via a buffer (implemented via a channel for now).

This allows the GUI to decide if it wants to synchronously wait for the
command to be processed or not. This is particularly evident for the
change name command (which has a deliberate 10 sec delay). The previous
code was forced (by the design of the FakeBus command dispatch) to wait
until the command had been processed. These changes allow the called to
decide if they want to wait or not, passing either a channel to receive
the result on, OR nil (if they want to plow ahead without waiting for
the command).

As far as I know, the amount of buffering in the `commandQueue` (
currently zero), titrates the blockingness of command submission. NB:
using a channel here should mean that commands are processed in the
order they are added to the channel, i.e. that their 'ordering' is
managed by the Golang channel concurrency management.