Update the hard-coded eventual consistency to be shorter
Changing the delay that Rename always has to be 2ms had the effect for
me of remaining obviously eventually consistent (i.e. a substantial
amount of time is always spent in the rename) AND being fast enough that
for my locally running copy, I would almost always see the change
already reflected in the read model after the POST handler for Rename
does the redirect to "/".